+++ src/main/java/duke/tasklist/Event.java
@@ -1,9 +1,12 @@
-package tasklist;
+package duke.tasklist;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
-
+/**
+ * Represent an event task in a list. It extends from the Task class and
+ * an Event object refers to an Event task which has description,date,start time, end time and the priority level.
+ */
 
 public class Event extends Task {
     protected LocalDate Date;
@@ -12,25 +15,15 @@ public class Event extends Task {
     private DateTimeFormatter displayTimeFormat = DateTimeFormatter.ofPattern("HH:mm");
     private DateTimeFormatter saveDateFormat = DateTimeFormatter.ofPattern("yyyy-MM-dd");
     private DateTimeFormatter saveTimeFormat = DateTimeFormatter.ofPattern("HH:mm");
-
-    public Event(String description, LocalDate Date, LocalTime startTime, LocalTime endTime){
-        super(description);
-        this.Date = Date;
-        this.startTime = startTime;
-        this.endTime = endTime;
-        super.isDone = false;
-        super.taskPriority = Priority.LOW;
-    }
-
-    public Event(String description, LocalDate Date, LocalTime startTime, LocalTime endTime, boolean isDone){
-        super(description);
-        this.Date = Date;
-        this.startTime = startTime;
-        this.endTime = endTime;
-        super.isDone = isDone;
-        super.taskPriority = Priority.LOW;
-    }
-
+    /**
+     * The event constructor creates an event object.
+     * @param description is the description of the Event
+     * @param Date is the Event Date
+     * @param startTime is the Event Start Time
+     * @param endTime is the Event End Time
+     * @param isDone is the Event task status
+     * @param taskPriority is the priority level for the task
+     */
     public Event(String description, LocalDate Date, LocalTime startTime, LocalTime endTime, boolean isDone, Priority taskPriority){
         super(description);
         this.Date = Date;
@@ -39,11 +32,21 @@ public class Event extends Task {
         super.isDone = isDone;
         super.taskPriority = taskPriority;
     }
+    /**
+     * The getDescription method overrides the method in Task class and
+     * returns the task in the following format to be represented on the screen
+     * [E][task status][priority level]Event description (Event Date Start Time - End Time)
+     */
 
     @Override
     public String getDescription() {
         return "[E]" + "[" + getStatusIcon() + "][" + super.getTaskPriorityToString() + "]" + super.getDescription() + " (at:" + Date.format(displayDateFormat) + " " + startTime.format(displayTimeFormat) + " - " + endTime.format(displayTimeFormat) + ")";
     }
+    /**
+     * The saveToFile method overrides the method in Task class and
+     * returns the task in the following format to be save to the file
+     * E | task status | priority level | task description | Event Date Start Time - End Time
+     */
     @Override
     public String saveToFile(){
         int taskStatus = isDone ? 1:0;
