--- src/main/java/tasklist/Priority.java
@@ -1,8 +1,0 @@
-package tasklist;
-
-public enum Priority {
-    HIGH,
-    MEDIUM,
-    LOW;
-
-}
